Animates a physio_timeplot, scales and adds to stimulus video into a composite movie

Animates a physio_timeplot, scales and adds to stimulus video into a composite movie

Usage

physio_movie(
  timeplot = NULL,
  stimfile = NULL,
  outputfile = NULL,
  fps = 30,
  spwidth = 800,
  spheight = 400,
  ...
)

Arguments

timeplot

a plot from function physio_timeplot, or pass variables for that function here and it will make a new one

stimfile

the stimulus video - leave blank and you can select it with the file OS

outputfile

name for composite movie. If blank, we'll just use the stim file name and add _spark

fps

frame per second of movie

spwidth

width of the sparkline movie - the sitmulus movie will be scaled to the same width before being composited

spheight

height of the sparkline movie

...

Arguments passed on to physio_timeplot

data

data list from physio_merge

plotmetric

The metric in data$dm to plot, eg DET or norm

plotdv

The dv, eg HR or EA

selection

A datatable with the columns and values that you want to use to select the data to plot

smoother

parameter for loess smoothing. set to 0 for none

plot_error

give error from smoothing